added main title for components
